FermiAppF
The last F stands for final.  I apologize for it being redundant.  The
changes include removing the teapots and changing the left handed
graphing position to right hand.  This was done by accessing the
MicroBooNE Unity file and changing Y rotation to 90.  Remember that the
Unity should be opened in Scene not Game.